% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wt_token.R
\name{get_wt_token}
\alias{get_wt_token}
\title{Download token using WildTrax password}
\usage{
get_wt_token(username)
}
\arguments{
\item{username}{String. Wildtrax username}
}
\value{
NULL. Does not return anything
}
\description{
Token set to WT_TOKEN
}
