% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_download_report.R
\name{wt_download_report}
\alias{wt_download_report}
\title{Download reports for ARUs from WildTrax}
\usage{
wt_download_report(
  project_ids,
  output_directory = NULL,
  include_birdnet = FALSE
)
}
\arguments{
\item{project_ids}{Vector of project ids to download}

\item{output_directory}{output directory to export to}

\item{include_birdnet}{Logical include birdnet outputs}
}
\description{
Exported to csvs
}
