% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_download_summary.R
\name{wt_download_summary}
\alias{wt_download_summary}
\title{Download summary of projects from WildTrax}
\usage{
wt_download_summary()
}
\value{
tibble of projects you can access
}
\description{
Only works for ARU data right now
}
